EarthPulse/backend/requirements.txt
# FastAPI Backend Dependencies for Earth Observation Visualizer
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
python-dotenv==1.0.0

# Geospatial Libraries for NASA EO Data
rasterio==1.3.9
geopandas==0.14.1
pyproj==3.6.1
shapely==2.0.2
fiona==1.9.5

# Data Processing
numpy==1.25.2
pandas==2.1.3
aiofiles==23.2.1
pillow==10.1.0
matplotlib==3.8.2

# Database (for future real data caching)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# Report Generation
reportlab==4.0.6
jinja2==3.1.2
openpyxl==3.1.2.0

# Optional: For enhanced NASA API integration
requests==2.31.0
beautifulsoup4==4.12.2
